% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cla}
\alias{cla}
\title{standard Laplace distribution}
\usage{
cla(a, xx, N)
}
\arguments{
\item{a}{variance}

\item{xx}{the initial value}

\item{N}{length of the chain}
}
\value{
a random walk of the standard Laplace distribution
}
\description{
Implement a random walk Metropolis sampler for generating the standard Laplace distribution
}
\examples{
\dontrun{
cla(1,0,2000)
}
}
