% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20089R.R
\name{n2kesplot}
\alias{n2kesplot}
\title{plot naive two demesions product kenel estimate}
\usage{
n2kesplot(x, h1, h2, N1, N2)
}
\arguments{
\item{x}{is sample data and x must be a n rows 2colums matrix}

\item{h1}{the bandwith of col1}

\item{h2}{the bandwith of col2}

\item{N1}{the width of estimate of col1}

\item{N2}{the width of estimate of col2}
}
\value{
a pic
}
\description{
plot naive two demesions product kenel estimate
}
\details{
This is some descriptio of this function.

plot the pic of two demesions product kenel estimate with specified bandwith and selected number of estimated points.
}
\examples{
 \dontrun{
n2kesplot(as.matrix(faithful),0.5,3,200,200)
}
}
