% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{nkes}
\alias{nkes}
\title{simple kenerl estimation function.}
\usage{
nkes(y, h, x)
}
\arguments{
\item{y}{is sample data}

\item{h}{bandwidth}

\item{x}{the point you need estimate}
}
\value{
density of x
}
\description{
caculate the density of the specified bandwith
}
\details{
This is some descriptio of this function.

the function can caculate the density of the specified bandwith
}
\examples{
 \dontrun{
nkes(faithful$eruptions,0.5,3)
}
}
