% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{nkesplot}
\alias{nkesplot}
\title{plot naive kenel estimate}
\usage{
nkesplot(y, N, h)
}
\arguments{
\item{y}{is sample data}

\item{N}{the number of estimate points}

\item{h}{bandwith}
}
\value{
a pic
}
\description{
plot the pic of naive kenel estimate
}
\details{
This is some descriptio of this function.

plot the pic of naive kenel estimate with specified bandwith and selected number of estimated points.
}
\examples{
 \dontrun{
nkesplot(faithful$eruptions,200,1)
}
}
