% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20089R.R
\name{pca}
\alias{pca}
\title{pca}
\usage{
pca(x)
}
\arguments{
\item{x}{is sample data matrix}
}
\value{
a pic and principal component and corresponding values
}
\description{
determin Principal Component
}
\details{
This is some descriptio of this function.

plot a pic of eigen value and return principal component and corresponding values
}
\examples{
 pca(state.x77)
}
